+++ packages/cli/src/command.ts
@@ -26,6 +26,7 @@ export const addCompileCommand = (parent = program, compiler?: Compiler): Comman
         .option("-d, --debug", "Enable the output of debug information.", false)
         .option("-p, --project <projectPath>", 'Path to the configuration file, or to a folder with a "tsconfig.json".', "./tsconfig.json")
         .option("-s, --sourceMap", "Enable the output of sourceMap information.", false)
+        .option("-t, --transpileOnly", "Enable the transpile only mode", undefined)
         .option(
             "-t, --targets <targetList>",
             "Comma-separated list of compilation target names to use specific configuration and list of addons.",
@@ -44,19 +45,6 @@ export const addCompileCommand = (parent = program, compiler?: Compiler): Comman
             }
         })
         .action((args: CompilerArguments, command: Command) => {
-            // TODO: Add support for style processors via addon
-            // process.stdout.write(`\nCompiling with configuration:\n`);
-            // process.stdout.write(`  + Config File: "${configPath}"\n`);
-            // let sassOptions: any = {};
-            // try {
-            //     const sassConfigPath = findSassConfig();
-            //     sassOptions = require(sassConfigPath);
-            //     process.stdout.write(`  + Sass Config: "${sassConfigPath}"\n`);
-            // } catch (error) {
-            //     process.stdout.write(`  + Sass Config: Cannot find module 'sass.config.js'. Using default config.\n`);
-            // }
-            // process.stdout.write(`\n\n`);
-
             // TODO: Add files from CLI argument
             const system = compiler?.getSystem() ?? compileSystem();
             const options = createOptions(args, compiler?.getReporter() ?? new DefaultReporter(system), system);