+++ packages/compiler/src/environment/index.ts
@@ -25,5 +25,4 @@ export {
     readFiles,
     recursiveFindByFilter,
 };
-
 export type { VersionedFile };